Move initialization dataflow impls into their own module.
